+++ junos/resource_security_nat_source.go
@@ -325,14 +325,14 @@ func setSecurityNatSource(d *schema.ResourceData, m interface{}, jnprSess *Netco
 	setPrefix := "set security nat source rule-set " + d.Get("name").(string)
 	for _, v := range d.Get("from").([]interface{}) {
 		from := v.(map[string]interface{})
-		for _, value := range from["value"].(*schema.Set).List() {
-			configSet = append(configSet, setPrefix+" from "+from["type"].(string)+" "+value.(string))
+		for _, value := range sortSetOfString(from["value"].(*schema.Set).List()) {
+			configSet = append(configSet, setPrefix+" from "+from["type"].(string)+" "+value)
 		}
 	}
 	for _, v := range d.Get("to").([]interface{}) {
 		to := v.(map[string]interface{})
-		for _, value := range to["value"].(*schema.Set).List() {
-			configSet = append(configSet, setPrefix+" to "+to["type"].(string)+" "+value.(string))
+		for _, value := range sortSetOfString(to["value"].(*schema.Set).List()) {
+			configSet = append(configSet, setPrefix+" to "+to["type"].(string)+" "+value)
 		}
 	}
 	for _, v := range d.Get("rule").([]interface{}) {