+++ junos/resource_security_nat_static.go
@@ -290,8 +290,8 @@ func setSecurityNatStatic(d *schema.ResourceData, m interface{}, jnprSess *Netco
 	setPrefix := "set security nat static rule-set " + d.Get("name").(string)
 	for _, v := range d.Get("from").([]interface{}) {
 		from := v.(map[string]interface{})
-		for _, value := range from["value"].(*schema.Set).List() {
-			configSet = append(configSet, setPrefix+" from "+from["type"].(string)+" "+value.(string))
+		for _, value := range sortSetOfString(from["value"].(*schema.Set).List()) {
+			configSet = append(configSet, setPrefix+" from "+from["type"].(string)+" "+value)
 		}
 	}
 	for _, v := range d.Get("rule").([]interface{}) {