+++ junos/resource_security_zone.go
@@ -470,9 +470,9 @@ func setSecurityZone(d *schema.ResourceData, m interface{}, jnprSess *NetconfObj
 		}
 		for _, v := range d.Get("address_book_set").([]interface{}) {
 			addressBookSet := v.(map[string]interface{})
-			for _, addressBookSetAddress := range addressBookSet["address"].(*schema.Set).List() {
+			for _, addressBookSetAddress := range sortSetOfString(addressBookSet["address"].(*schema.Set).List()) {
 				configSet = append(configSet, setPrefix+" address-book address-set "+addressBookSet["name"].(string)+
-					" address "+addressBookSetAddress.(string))
+					" address "+addressBookSetAddress)
 			}
 			if v2 := addressBookSet["description"].(string); v2 != "" {
 				configSet = append(configSet, setPrefix+" address-book address-set "+